+++ bootique-swagger-openapi/src/test/java/io/bootique/swagger/openapi/SwaggerOpenapiModuleProviderTest.java
@@ -17,38 +17,32 @@
  * under the License.
  */
 
-package io.bootique.swagger.ui;
+package io.bootique.swagger.openapi;
 
+import io.bootique.BQRuntime;
 import io.bootique.jersey.JerseyModule;
-import io.bootique.swagger.SwaggerModuleProvider;
+import io.bootique.test.junit.BQModuleProviderChecker;
+import io.bootique.test.junit.BQRuntimeChecker;
 import io.bootique.test.junit.BQTestFactory;
-import org.junit.BeforeClass;
-import org.junit.ClassRule;
+import org.junit.Rule;
 import org.junit.Test;
 
-import javax.ws.rs.core.Response;
+public class SwaggerOpenapiModuleProviderTest {
 
-import static org.junit.Assert.assertEquals;
+    @Rule
+    public BQTestFactory testFactory = new BQTestFactory();
 
-public class SwaggerUiDefaultLocationIT extends SwaggerUiBaseIT {
-
-    @ClassRule
-    public static BQTestFactory TEST_FACTORY = new BQTestFactory();
-
-    @BeforeClass
-    public static void beforeClass() {
-        TEST_FACTORY.app("-s")
-                .module(new SwaggerModuleProvider())
-                .module(new SwaggerUiModuleProvider())
-                .module(b -> JerseyModule.extend(b).addResource(TestApi.class))
-                .run();
+    @Test
+    public void testAutoLoadable() {
+        BQModuleProviderChecker.testAutoLoadable(SwaggerOpenapiModuleProvider.class);
     }
 
     @Test
-    public void testApi_Console() {
-        Response r = BASE_TARGET.path("/swagger-ui").request().get();
-        assertEquals(200, r.getStatus());
-        assertEqualsToResourceContents("swagger-response1.html", r.readEntity(String.class));
+    public void testModuleDeclaresDependencies() {
+        final BQRuntime bqRuntime = testFactory.app().module(new SwaggerOpenapiModuleProvider()).createRuntime();
+        BQRuntimeChecker.testModulesLoaded(bqRuntime,
+                JerseyModule.class,
+                SwaggerOpenapiModule.class
+        );
     }
-
 }