+++ RELEASE-NOTES.md
@@ -1,3 +1,11 @@
+## 2.0
+
+* #16 Support for OpenAPI 3.0.x
+* #17 Configurable Swagger UI servlet mapping 
+* #18 Change default mapping for UI servlet from /swagger to /swagger-ui
+
+## 1.0
+
 ## 1.0.RC1
 
 * #1 Add integration with Swagger web console